rename components and update imports.
